WIP: Support all joysticks, and detaching sticks
This commit changes the WIP joystick API to have a separate Observable
for the axes of each stick supported by GLFW.  Each Observable's value
may be either a Vector{Float64} for the axis values, or Nothing if the
stick is not attached.  By allowing Nothing values, Makie again works if
there's no joystick plugged in, which is a plus. :)